// packages/core/src/types/types.ts
export interface Link<T extends string> {
  sys: {
    type: 'Link';
    linkType: T;
    id: string;
  };
}
// ...
export type DataSourceEntryValueType = Link<'Entry' | 'Asset'>;
// ...
export type CompositionNode = {
  definitionId: string;
  children: Array<CompositionNode>;
  variables: Record<string, CompositionComponentPropValue>;
};

export type CompositionDataSource = Record<string, DataSourceEntryValueType>;
export type CompositionUnboundValues = Record<string, { value: CompositionVariableValueType }>;

export type Breakpoint = {
  id: string;
  query: string;
  displayName: string;
  previewSize: string;
};

export type SchemaVersions = '2023-09-28' | '2023-06-27' | '2023-07-26' | '2023-08-23';

export type ExperienceComponentSettings = {
  variableDefinitions: Record<
    string,
    Omit<ComponentDefinitionVariableBase<ComponentDefinitionVariableType>, 'defaultValue'> & {
      defaultValue: CompositionComponentPropValue<'BoundValue' | 'UnboundValue'>;
    }
  >;
};

export type Composition = {
  title: string;
  slug: string;
  componentTree: {
    breakpoints: Array<Breakpoint>;
    children: Array<CompositionNode>;
    schemaVersion: SchemaVersions;
  };
  dataSource: CompositionDataSource;
  unboundValues: CompositionUnboundValues;
  usedComponents?: Array<Link<'Entry'> | ExperienceEntry>;
  componentSettings?: ExperienceComponentSettings;
};

export type ExperienceEntry = {
  sys: Entry['sys'];
  fields: Composition;
  metadata: Entry['metadata'];
};